+++ page_object/pages/feed_page.py
@@ -28,18 +28,22 @@ class FeedPage(BasePage):
         return method, locator
 
     @allure.step("Найти заказ по номеру")
-    def find_order_by_number(self, number, page_locator):
-        order_locator = self.calculate_locator(number, page_locator)
+    def find_order_by_number(self, number):
+        order_locator = self.calculate_order_locator(number)
         order = self.find_element_without_wait(order_locator)
         return order
 
-    @allure.step("Найти заказ по номеру")
-    def set_order_locator(self):
-        return FeedPageLocators.ORDER_BY_NUMBER
+    @allure.step("Вычислить локатор заказа на странице")
+    def calculate_order_locator(self, order_number):
+        order_locator = self.calculate_locator(order_number, FeedPageLocators.ORDER_BY_NUMBER)
+        return order_locator
 
     @allure.step("Найти Popup с информацией по заказу")
     def find_popup_order_details(self):
         return self.find_element_with_wait(FeedPageLocators.MODAL_ORDER_FORM)
 
-
+    @allure.step('Найти текст заголовка ленты заказов')
+    def get_feed_header_element(self):
+        element = self.find_element_with_wait(FeedPageLocators.FEED_HEADER_TEXT)
+        return element.text
