+++ tests/test_navigation.py
@@ -2,6 +2,7 @@ import allure
 import pytest
 
 from page_object.pages.constructor_page import ConstructorPage
+from page_object.pages.feed_page import FeedPage
 from page_object.pages.headers_page import HeadersPage
 
 
@@ -19,7 +20,7 @@ class TestNavigation:
 
         constructor_page.open_main_page()
         headers_page.click_on_constructor()
-        assert headers_page.find_constructor_burger_text()
+        assert constructor_page.find_constructor_burger_text()
 
     @allure.title('Переход с главной страницы на страницу Лента заказов по клику на надпись Лента заказов')
     @allure.description('Пользователь переходит на главную страницу, кликает на надпись Лента заказов и '
@@ -30,9 +31,10 @@ class TestNavigation:
         driver = request.getfixturevalue(br_driver)
         constructor_page = ConstructorPage(driver)
         headers_page = HeadersPage(driver)
+        feed_page = FeedPage(driver)
 
         constructor_page.open_main_page()
         headers_page.click_on_feed()
-        feed_header = headers_page.get_feed_header_element()
+        feed_header = feed_page.get_feed_header_element()
         assert feed_header == "Лента заказов"
